Nav improvements for variable font sizes
Nav set to turn into a column on smaller screen sizes to provide more flexibility when user uses browser features to increase or decrease font size.

Column nav has been provided in lieu of a collapsed menu option as this navigation serves as more of a table of contents, rather than a navigation for a multi-page site.